% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_titanc.R
\name{calculate_survival_rate_by_class}
\alias{calculate_survival_rate_by_class}
\title{Calculate Survival Rate by Class}
\usage{
calculate_survival_rate_by_class(data)
}
\arguments{
\item{titanic_data}{A data frame containing the Titanic passenger data.}
}
\value{
A data frame with the total passengers, survived passengers, and survival rate for each class.
}
\description{
This function calculates the survival rate for each passenger class on the Titanic.
}
