% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_titanic.R .r
\name{taux_survie_classe}
\alias{taux_survie_classe}
\title{Cette fonction calcule la proportion de survivants pour chaque classe.}
\usage{
taux_survie_classe(data)
}
\arguments{
\item{data}{Jeu de données TitanicSurvival}
}
\value{
Un tableau avec la proportion de survie par classe.
}
\description{
Cette fonction calcule la proportion de survivants pour chaque classe.
}
