% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_titanic.R .r
\name{taux_survie_sexe}
\alias{taux_survie_sexe}
\title{Cette fonction calcule la proportion de survivants pour chaque sexe.}
\usage{
taux_survie_sexe(data)
}
\arguments{
\item{data}{Jeu de données TitanicSurvival}
}
\value{
Un tableau avec la proportion de survie par sexe.
}
\description{
Cette fonction calcule la proportion de survivants pour chaque sexe.
}
