% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TD_reprod.R
\name{calculate_survival_rate_by_sex}
\alias{calculate_survival_rate_by_sex}
\title{Calculate survival rate by sex}
\usage{
calculate_survival_rate_by_sex(data)
}
\arguments{
\item{data}{Un data frame contenant les informations du Titanic (doit inclure les colonnes 'sex' et 'survived').}
}
\value{
Un data frame avec les taux de survie calculés pour chaque sexe.
}
\description{
Cette fonction calcule le taux de survie en fonction du sexe des passagers du Titanic.
Elle prend un jeu de données et retourne un tableau contenant les taux de survie pour chaque sexe.
}
